Remove old LoginViewController and rename LoginNewViewController to take its place
